requirements.txt
adrf==0.1.8
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-property==0.2.2
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.4
django-redis==5.4.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
frozenlist==1.5.0
h11==0.14.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
magic-filter==1.0.12
multidict==6.1.0
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
uvicorn==0.32.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
